cast to size_t to get rid of warning

+++ plane_seg/include/plane_seg/common/LoadData.h
@@ -134,7 +134,7 @@ inline void loadEigenMatrix(const std::string& filename, const std::string& matr
     }
   }
 
-  if (numFailed == matrix.size()) {
+  if (numFailed == static_cast<size_t>(matrix.size())) {
     throw std::runtime_error("[loadEigenMatrix] Could not load matrix \"" + matrixName + "\" from file \"" + filename + "\".");
   } else if (numFailed > 0) {
     std::cerr << "WARNING: Loaded at least one default value in matrix: \"" + matrixName + "\"\n";
